Added SSL/password sidecar
Added a sidecar for handling secrets related to database login. For
logging in, the sidecar populates a home directory which can be mounted
in other containers. This makes it so that apps don't have to retrieve
their own db secrets, and also makes database connection strings
simpler.

Next, propagate the change to all apps.